requirements.txt
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
xgboost>=1.4.0
lightgbm>=3.2.0
catboost>=0.24.4
matplotlib>=3.3.0
seaborn>=0.11.0
shap>=0.39.0
imbalanced-learn>=0.8.0
mlflow>=1.20.0
optuna>=2.8.0
hyperopt>=0.2.5
pyyaml>=5.4.0
category_encoders>=2.3.0